src/azure-ai-102/language/02-language-agent/pyproject.toml
[project]
name = "text-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = ['python-dotenv', 'azure-identity', 'azure-ai-projects==2.0.0b4']
